Wishes are now called "action ideas".
I will probably bring back the term "wishes" later to mean,
specifically, things that people want brought to them, like "a puppy",
"a latte", etc, or personal experiences people want to have, like "to
ride a pony".

But wishes that are specifically organizing ideas will be called action
ideas for now.  It is blander but less gay.